pyproject.toml
[tool.poetry]
name = "oh-my-goose"
version = "0.1.0"
description = "Orchestration layer on top of Goose, inspired by oh-my-claudecode"
authors = ["Adam McIntosh <adam@adam-mcintosh.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "oh_my_goose"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
pydantic = "^2.5.0"
rich = "^13.7.0"
textual = "^0.47.0"
watchdog = "^3.0.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
pytest = "^7.4.3"

[tool.poetry.scripts]
omg = "oh_my_goose.cli.omg:main"  # Adjust if your CLI entrypoint is in cli/omg.py

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"